requirements.txt
streamlit
langchain
langchain-community
streamlit-chat
python-dotenv
openai
langchain_openai
pypdf
pandas
python-docx
langchain_text_splitters
langchain-huggingface
faiss-cpu
transformers
huggingface_hub
langchain-chroma
bs4
wikipedia
youtube_search
arxiv
langchainhub
yfinance
FastAPI
jinja2
